+++ src/view/film-cart/film-card.js
