+++ src/view/films-count/films-count.js
