+++ src/view/films/films-list-top-rates.js
@@ -1,0 +1,8 @@
+import {createFilmTitle} from './films-title';
+
+export const createFilmsListTopRates = () => (
+  `<section class='films-list films-list--extra'>
+      ${createFilmTitle('Top rated')}
+      <div class='films-list__container' id='topRatesList'></div>
+    </section>`
+);
